fix: volume slider on android

+++ unisync_android/app/src/main/java/com/anhquan/unisync/ui/screen/home/HomeViewModel.kt
@@ -42,7 +42,7 @@ class HomeViewModel : ViewModel() {
         _state.update { state ->
             state.copy(
                 status = Status.Loaded,
-                volume = device.getPlugin(VolumePlugin::class.java).currentVolume.toFloat()
+                volume = device.getPlugin(VolumePlugin::class.java).currentVolume
             )
         }
     }
